refactor: extract canvas type definitions from canvasStorage.ts
Move GroupActivation, ControllerOption, UnitVariant, ControllerDef, GroupDef,
and CanvasDocument interfaces into canvasTypes.ts so pure logic can be unit
tested without React mocking. canvasStorage.ts re-exports all types for
backwards compat; no downstream imports need changes.
